Roll back to REST API endpoints
